feat: prevent gaining experience in the Abyss
Abyss scumming has been present in the game for a long time, a lingering
design wart where the running thinking is that the risks outweigh the
benefits. With the changes to the Hells and Tomb this wisdom is a bit
lackluster; indeed the strategic damage of the Hells can be countered by
"working off" the stat and HP drain in the abyss.

This commit locks out progress towards all XP-gated effects except the
abyss progress timer.