feat: remove abyss random item generation
The second half of abyss scumming, collecting consumables on Abyss:1 and
departing promptly on the first offered XP gate exit. Again, this has
historically been regarded as not-worth-the-risk, but the shifting
nature of the end game means that balance has changed.

Vaults can still place items, but vaults are scary and sufficiently rare
that this should be low enough percentage to be not a good idea to scum.