Gloves of Parrying
The anti-acrobat, they provide SH bonus if the last action was a melee attack.
This bonus is halved if your offhand is occupied by a non-weapon item.

(People keep asking for two handers to have innate SH, which seems like bad
design, but there may be space for a limited availability ego in that idea).