Change Amulet of Alchemy's secondary effect
Instead of reducing the cost of Alchemy spells by 1, it now restores a
small amount of MP (currently 3-6) whenever the wearer drinks an
otherwise-useful potion.

This is a less narrow secondary effect than before (it seems a good idea to
me for amulets to be a little broader in their appeal) and leans into
another thematic association of alchemy. Additionally, MPRegen amulets were
often by-far the most desireable amulet for many mages and this might give
it a little competition in that regard (ideally being a little less
strong passively, but better in burst situations).

To discourage fiddly amulet swapping, it requires attuning to the amulet at
full MP, just like MPRegen does.

(It's possible I've set the numbers too high, but I'm curious how players
will feel to use it.)