Slightly change item hotkey behavior in 'i'nventory menu
Pressing a letter now jumps to the first item *below* the cursor (wrapping
around, if necessary) rather than 'nearest'.

This interacts better with using category hotkeys, like ?, to jump between
categories, since the item you want will always be below that point, but
may not be nearest that point.

(And it turns out, the comment already claimed it did this, from when it
used to do this during development.)