Update to Android SDK 35
Starting August 31 2025, new apps and app updates must target Android
15 (API level 35) or higher to be submitted to Google Play.

There are two changes related to window insets in Android 15:
edge-to-edge is enforced by default, and there are also configuration
changes, such as the default configuration of system bars.

I'm also reworking the code to control the virtual keyboard
visibility because it's a mess.