Further adjust Kiku's wrath (Oneirical)
When b6bfbb2 reduced Kikubaaqudgha's draining wrath effect (from a range
between ~23.3% to 40% max hp to a new range of ~16.6% to 30% max hp) and
removed the ability to resist it via rN simultaneously (so an rN+++
character couldn't ignore a large part of the wrath), it made the wrath
still capable of dragging former worshippers down to single-digit percents
of their max HP (through the randomness of when the wrath actually
triggers), but without very much potential recourse to handle such.
Gifting god wrath should be dangerous, but even Okawaru wrath summons have
more potential responses than this, so I'm further adjusting the numbers
and adding another angle.

The new Kiku zero-tension wrath effect is mildly reduced at the top end,
with a range of ~16.6% to ~26.6% max HP, and more importantly will not
apply at all if the player is already drained for 50% or more of their max
HP. In exchange, Kiku wrath now also directly inflicts 35% to 50% Doom
regardless of whether or not draining occurred, as is somewhat linked to
the undead between draugr attacks and mummy death curses. Banes are still
reasonably rare to encounter currently (without entering optional /
extended branches), are evidently currently reasonably survivable, and
add to the strategic-damage emphasis prior; as such, it should be safe to
emphasize these over the relatively conventional draining angle.