Don't trigger Stardust by using Sanctuary (Lici)
Not likely to come up much in a real scenario, since it would need to be
your first use of MP in a given battle, and still doesn't result in penance
if enemies are near you (since Zin suppresses all seeker hits inside a
sanctuary).

But technically, if there's only monsters *outside* the sanctuary, they
can fly out and hit *those* and give you penance. So let's just avoid that.