Don't consider books of evil spells to be themselves evil (salticid)
So that autoexplore will not ignore them. Swapping away from good gods is
done more often than other gods, and the current behavior means that if a
player wants to use any evil spell they already encountered, they need to
manually go back and pick it up.

The spell library is supposed to remove this need in general, and it seems
harmless for forbidden magics to sit there in case the player may be
tempted to sin in future.