Don't wait until the player is in LoS to regenerate Nobody's memories
When a large amount of time passes, Nobody will recover enough memories to
fully regenerate, but this wasn't being done until the player was in LoS
(and Nobody was aware of them), which could briefly result the game
claiming they had few/no memories left and then appearing to spontaneously
generate them as soon as they noticed you. Now, it should restore itself
in a more player-indifferent fashion.

Also fix a minor typo.

This fixes #4673