Fix some orb of pyromania timing issues (Acrobat)
Pyromania intentionally could trigger only once per turn, but it did this
by storing the number of qualifying kills and then rolling the chance at
the end of the player's turn.

However, the player can kill things on the enemy's turn - this is how
things like clouds, poison, noxious bog, and several other things work.
Which resulted in an explosion at the end of the turn *following* the
monster dying.

To improve clarity, the explosion is now triggered immediately when a
qualifying monster dies, using an internal cooldown that only resets at the
start of the player turn. This should hopefully work much better.