Don't dismiss eldritch tentacles on player attacks (Acrobat)
The previous behavior of 'getting angry at the player' was seemingly broken
by fb19e97 and unnoticed by me until now. More than most summons, the fact
that you may have to deal with a hostile monster in a random position is a
deliberate part of this spell and being able to avoid it by hitting the
tentacle first (perhaps to 'reroll' its placement), feels like incorrect
behavior. (Although, in most cases, this is unlikey to matter much.)