Give Gozag's Platinum Reserve a spit-shine
Gozag's Platinum Reserve was created mostly as a joke: a shop full of
expensive artefacts you can't afford, intended to remind you that you're
poor -- but you can change that by joining Gozag! It predates my modern
philosophy of "a shop that's always bad is a joke that's only funny
once".

In practice, it's not always so bad -- sometimes, players will
accumulate a ton of gold in late game or extended and need something to
spend it on. But the problem with Gozag's Platinum Reserve was the
choice -- it only generated four items, hardly enough selection for the
discerning consumer!

Combining the notion of Gozagian capitalist excess with a shop gimmick
I've been thinking about for a while, I have touched up the inventory
by making it generate twenty items: the maximum a shop can hold. I've
also made sure to include magical staves and talisman artefacts, which
did not exist when the Reserve was initially created. Now you're more
likely to find something that's still worth it even at 10x the price.
Or, at the very least, you'll find something that makes you really wish
you weren't poor, and you'll look over at the Gozag altar...

Sadly, it's still not possible to add the biggest laser, Airwolf, or the
head of Keith Moon to shop inventories, but I did add some golden
statues to the grounds to punch things up a little more.