Allow transporters to count for level connectivity checks (regret-index)
The part of map generation that verifies all critical parts of a floor are
reachable (such as branch entrances) did not understand transporters (and
in fact was performed before transporters were properly set up, so it
couldn't see them even if it tried).

It seems to me like these should count for connectivity the same as
anywhere else the player can walk, so now we set up transporters slightly
earlier in level generation and allow connectivity floodfill to traverse
them in a similar way that travel already did.