Try changing Boundless Tesseract's activation condition, tweak slightly
While overall reaction to tesseracts, now that the dust has settled a bit,
seems to be moderately positive, but one pain point in particular stands
out: when the tesseracts actually activate.

Currently they respond to noise and stealth checks the same as any other
monster (even if many people has ascribed custom behavior to them), and
this works okay in *most* situations, but there's some opacity in when
noise will actually reach them (ie: chains of monsters shouting can awaken
it from further away than expected, and it may not be clear why this
happens sometimes but not others), but also they will reliably get woken up
by alarm traps (even from exploration - possibly as soon as one enters the
floor!) and sometimes weirder situations like a passively made Jiyva jelly
bumping into one on the other side of the level. In general, this can feel
a bit arbitrary and unsatisfying.

So instead, this is an attempt to link the activation to something directly
observable by the player and more transparent, while hopefully also not
being gameable in as many ways as some other proposals: tesseracts now
activate the first time an orb guardian is killed (and orb guardians no
longer follow the player off Zot:5 unless the orb run has been started, so
that the player is not *forced* to activate the tesseract on an unlucky
Zot:5 entrance)

While it is technically possible to lure them away from the orb vault to
another staircase, the fact that they are all patrolling (and fast) means
that it's nearly impossible to 'cheese' anything in this way. Early alarm
traps can still be nasty, but the player may have options to avoid
triggering the tesseracts (and the orb guardians in the initial chamber
have been nudged backwards to make them less likely to pour out
immediately.)

In addition, since I expect the distance the player is from the tesseract
when it activates to be shorter in the average case (and due to certain
feedback about throwing/ranged weapons), I have moved the tesseracts one
tile around the corner into the upper lung. This is far more gentle than
some suggestions about giving them warding, since it is still possible to
damage them at range from the back of the lower lungs (and with large AoE
spells), but will require the player to expose themselves to a bit more
danger in the process.

Finally, the total number of tesseract-made spawns on the floor is now
capped at 100 (already rather out of control, but likely sufficiently bad
that we don't need to make outliers even more absurd), the maximum spawn
rate is slightly reduced, and it takes slightly more turns to reach it
(1200 instead of 1000).