Attempt to improve the Orb card again
The 'starburst' effect from the Orb card was always more flashy than
effective - none of the orbs were launched with respect to the player's
chosen aim, and yet it was overly aggressive about telling orbs to attempt
to home in on that target anyway (instead of targets they might have a
chance of actually hitting). This meant that 6 orbs were often less good
at hitting anything at all than 1 orb.

When f4014ab attempted to fix a seperate issue with the starburst being the
'low' power effect of the card, it greatly exacerbated this problem; not
only did the single-orb version become much rarer at high power, but the
starburst effect could now fire just 1 or 2 orbs - usually aimed at nothing
in particular - making it dramatically ineffective.

This commit makes the following changes:
 -The 'center' of the starburst effect is always aligned with the player's
  aim (so that one orb always follows the direct path)
 -Non-center orbs will now choose targets based on 'closest angle' rather
  than 'closest distance within a moderate angle'. In general, this makes
  them a lot more likely to hit something (or at least make a reasonable
  attempt at doing so).
 -Burst orbs that lose their target lock (often due to another orb killing
  their target first) will try to reacquire the best one for their current
  angle and position.
 -Card power 1 always fires 3 orbs (in a narrower spread), while power 2
  fires 6 in a wider spread.

In general, higher power effects should now almost always be strictly
better than low power ones (instead of worse) and considerably more
effective. One downside is that the new target lock behavior does make the
orbs more likely to perform a full 180 and return in the player's
direction (provided an enemy is in the way), though one can usually still
safely maneuver them into hitting that enemy in this circumstance (and
surely a little bit of danger is appropriate for the gambler god -
especially when the payoff is large).