Greatly reduce the noise of ioods hitting each other
It's not at all clear why this was *much louder than Shatter* (and, in
fact, tied for the loudest noise in the game, along with things like Gong
and alarm traps....)

Orb/orb collisions are a lot more likely with the revised Orb card than
they used to be (though not all bad - since they will often be next to an
enemy when they collide!), and that level of noise seemed unexpected and
undesireable. They're still very loud, but no louder than Storm card.