Fix monster airstrike damage flagrantly lying in xv
The listed damage for monster airstrike was extremely inaccurate in
multiple ways, treating it as though it used the same formula as player
airstrike when it very much did not. For instance, wind drakes claimed that
they did 2d(7-23) damage, but their actual damage range with 0 empty spaces
was 0-15 and with 8 empty spaces was 16-31. This is a very similar average,
but much lower variance (and a maximum damage that is a whole 33% lower
than claimed!)

This is a result of:
 -Using a function of random2avg() instead of roll_dice() for the
  power-based componant, which is extremely similar but not identical to
  the displayed numbers.
 -Not randomizing the space-based componant at all. You just flatly take
  +2 damage per empty space (unlike the player version).

While I considered changing the formula to match the player version, I feel
that the difference between how the open space bonus is handled makes
mechanical sense: the player is more predictably punished for being out of
position, while not being able to so consistently punish enemies for the
same 'mistake' (since enemies will make no attempt to avoid this). Also,
fixing it would mean giving multiple enemies significantly higher max
damage (which is something past commits already worried out loud about).

It's not clear that Gastronok needs to hit for 16 more damage than he
currently does, so I am instead opting to change xv to convey the true
behavior about space, while leaving it unchanged. (I did replace the use of
random2avg with roll_dice, since that difference is too unimportant not to
make consistent).

(This commit also fixes a minor bug where the player was considered 'open
space' for purposes of a monster airstriking something adjacent to you.)