Highlight V:$ and Zot:$ staircases (Darby)
The guaranteed Vaults end ambush is iconic and unique and mechanically
quite successful at forcing a different style of play to approach the
floor. This arrangement is also completely unforecasted for both unspoiled
newcomers and experienced players playing fast enough they forgot which
floor of Vaults they're on. As such, this commit is adding unique stair
tiles and descriptions to the stairs leading in and out of Vaults:5, which
should both be a nice visual highlight late in the game as well as help
warn the aforementioned players from entering such.

(While it's unlikely people will miss they're entering the last floor of
Zot, due to the fixed colours of each floor, it's nice to have more tile
highlights for the climactic finish to a standard game, so the staircases
leading in and out of Zot:5 have also gotten new tiles. They have not
gotten unique descriptions, though, since there's no guarantees to what
one sees on entering the floor.)

All of these tiles were textures made by Darby on top of white_noise and
enne's edits of Denzi's staircase tiles. Said descriptions were also
written by Darby.