Don't pre-reveal transporter destinations in Necropolis/Gauntlet
Travel can be overzealous in taking these transporters when players don't
expect it to, resulting in dumping the player into extremely scary
situations without warning.

Some better solution would involve displaying transporter destinations but
still not pathing the player through them if doing so would be dangerous,
but in the meantime the lesser evil is certainly the one that doesn't kill
people.