Add baubles to "autopickup toggle" menu
(Note: the autopickup toggle menu is primarily/internally intended for
"known items" in comparison to unknown items)

Baubles were previously not in this menu, and it seemed logical to add
them. In particular, the entire "Baubles" category is added,
futureproofing if other baubles besides flux are created.

Closes #4619.

[Committer's note: Moved the bauble check in check_item_knowledge() to
just after the talisman check, removed sorting of baubles to match
talismans (since added bauble types are likely to have tiers
eventually), and adjusted a comment regarding the latter.]