Handle swamp_mist_pool teleport closets (#4735)
The chances of getting a teleport closet here when the vault is pressed
up against a wall are slim (since surrounding tiles also have to roll for
trees instead of water), but not impossible. As such, I am using two
different ways to address the matter in a randomized fashion: forced level
connection points (which doesn't always make corridors in many open
layouts, but does it enough times I'm avoiding it being the default), and
using the CLEAR syntax (which then relies on the default level connection
checks). While I'm here, I also cleaned up the header a little.

Closes #4735.