Fix a possible quick blade crash
If a coglin killed a monster with their first weapon, there would be a null
defender for the second weapon swing - which normally isn't an issue, since
this is used deliberately to setup possible cleaving on that weapon and
bails out before it actually tries to *attack* that null defender.

But if the second weapon a quick blade, and its first attack fumbled in
water, it could try to perform a second attack against that defender
without first checking if it was valid, resulting in a crash.