Adjust monster door-smashing behavior (various)
While monsters smashing closed doors was intended to stop the weird dance
of the player opening a door as rapidly as another monster closes it (to
regenerate while staying safe) and did fairly well in that capacity, due
to specifics of monster behavior code, a monster would never become
eligable to smash a door unless they had physically seen the player before.
Some configurations of door allowed the player to do this indefinitely.

Instead, require the monster be targeting the player and the player be
adjacent to the door (otherwise, they can't close it again anyway!). This
should solve the potentially degenerate case while preventing a swath of
broken doors from wandering monsters elsewhere on the level.

(But berserking monsters can just break doors whether you're near them or
not.)