Fix a crash with examining submerged shops (Omelet, Darby)
If the player had seen a shop on a given tile (or had it pre-revealed to
them in a Necropolis), but that tile was not currently in view and had
also had its terrain changed (such as by a nearby water nymph) examining
that tile (or even just moving your mouse over it in local tiles) would
crash.

This is a little bit of a cludgy fix (and still imperfect, since trying to
examine such a shop will show a question mark instead of its normal icon),
but a lot of code calls shop_at as a way of checking if there *is* a shop
at some arbitrary space, rather than in situations where it 'knows' there
is one, and this avoids disrupting any of that.