Remove undefined equipment behavior when transiting monsters (Xolotl)
In general, monster inventories are not stored inside the monster
themselves, but in env.items and the monster just stores references to
indicies in that array that correspond to their equipment.

When monsters are transited from one floor to another, their old inventory
is copied and new links set up when they are placed. *However*, determining
where to put them has to check if the monster is flying, which has to check
their equipment, *which isn't set up yet, and is thus pointing to random
items on the new floor*. In most cases, this is invisible, since most items
are either empty or have no relevant properties, but in theory a monster
*could* incorrectly think it was flying if one of its links pointed to a
flight item on the new floor.

Or, as led to this being discovered, could crash if the monster thought it
was wearing an artefact spellbook.

I believe this bug has probably existed approximatetly forever, but just
been very hard to trigger in a relevant way.