Let autofight move towards warded enemies (various)
Autofight understood that it was impossible to damage warded enemies while
at range and would not consider attacking them. This is desireable for
ranged combat (so that it will shoot other things instead), but more
awkward for melee characters or when there is only one enemy in sight.

Now autofight should move towards a warded enemy if there are no other
valid targets the player can currently damage.

(This will be even more important in the near future when warded enemies
will exist beyond one bane and one wizlab.)