De-spoiler throwable mulch rates (Ge0ff)
Instead of saying "It may be destroyed on impact," state the chance of
destruction on impact as a fraction (as many of the percentages are non-
integers or less visibly round). Refactors to get the brand-modified
mulch rate in a separate function from the actual check to mulch, so it
can be displayed to the player without code duplication.