Remove some (hopefully) needless special-casing of webtiles status export
I cannot find any current reason these statuses are special-cased (one of
them didn't even display properly in webtiles until the previous commit).

Looking through the history, there is some suggestion acrobat was done this
way to make the timing of it turning blue work in webtiles, but it
currently seems to have no effect on that, either way.