Move mark of execution's rampage check into player::rampaging
This check was added to the wrong method in a31ef7ce32.

There is no existing implementation of monster rampage, so this didn't
matter for now, but if that ever changes it would be good to have this
check in the right place so that all actors don't start rampaging while
the player's mark of execution is active.