Bring upstream packages closer to Debian/Ubuntu distributed packages
Hi, I'm the new maintainer for crawl in Debian, and I'd like to sync up
the upstream Debian packaging with what's currently released in [Debian
sid](https://packages.debian.org/source/sid/crawl). I merged a handful
of changes from the upstream packaging into the Debian version (thanks
to @robertxgray in particular for the changes to use
`install-xdg-data`), and I'd like to contribute back some of the changes
I made. The goal here is to reduce the diff as much as I can (outside of
debian/changelog) to make it easier for me to package newer versions of
crawl in Debian going forwards, and also to reduce the likelihood that
the Debian/Ubuntu vs upstream packages diverge enough to cause problems
e.g. APT upgrade conflicts, although that's already highly unlikely.
Thanks for considering!