Bump most CI to ubuntu-latest, python to 3.13
Here we bump python versions to 3.13 so that we can test webtiles with
crypt-r, and we bump as many runners as we can back up to ubuntu-latest.

Webtiles runners now test with python 3.8, 3.11, and 3.13, with older
versions using crypt, and 3.13 using crypt-r following 7c3bd3c27620.

We still have some issues with the mingw, catch2, and android runners
under ubuntu-latest, so these runs remain using ubuntu-22.04 until we
can get those sorted.