Fix scrolls being wasted when losing connection (mumra)
When reading an an unidentified scroll that requires you to make a
choose such as enchant weapon that requires you to choose a weapon, if
you save and quit before making the choose the scroll would be wasted.
The game can save and quit on its on when losing connection in webtiles
resulting in the player losing a scroll at no fault of their own. To fix
this restore the choose on loading the game like we do with other
uncancelable chooses.