refactor cloud_tile_variation
cloud_tile_variation is an enum that determines how a cloud chooses its
tile. It has an entry called CTVARY_NONE which is used for clouds with a
fixed tile that doesn't vary. However, some clouds with it were special
cased to not have a fixed tile, which adds extra complexity over givng
them their own enum entries and also makes it harder to determine which
clouds are animated.