Use DejaVu Sans font by default in webtiles
The adds the font as an embedded webfont, so webtiles now appears the
same as the default desktop font.

The font is still overridable in configuration should anybody prefer to
use their own system or other font, but this certainly looks much nicer
than the default monospace font on browsers in Windows!